src/pyproject.toml
[project]
name = "multi-agents-cykg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hf-xet>=1.1.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.12",
    "langchain-huggingface>=0.3.1",
    "langchain-neo4j>=0.5.0",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "mcp-use>=1.3.7",
    "numpy>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rdflib>=7.1.4",
    "sentence-transformers>=5.0.0",
]
